Extracted the console output view into its own FXML & controller
