+++ scs2-session-visualizer-jfx/src/main/java/us/ihmc/scs2/sessionVisualizer/jfx/session/mcap/MCAPLogSessionManagerController.java
@@ -12,15 +12,12 @@ import javafx.fxml.FXML;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
-import javafx.scene.control.ListView;
 import javafx.scene.control.TextField;
 import javafx.scene.control.TextFormatter;
 import javafx.scene.control.TitledPane;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.FlowPane;
-import javafx.scene.paint.Color;
-import javafx.scene.text.Font;
 import javafx.stage.FileChooser;
 import javafx.stage.FileChooser.ExtensionFilter;
 import javafx.stage.Stage;
@@ -32,8 +29,6 @@ import us.ihmc.log.LogTools;
 import us.ihmc.messager.TopicListener;
 import us.ihmc.messager.javafx.JavaFXMessager;
 import us.ihmc.scs2.session.SessionRobotDefinitionListChange;
-import us.ihmc.scs2.session.mcap.MCAPConsoleLogManager.MCAPConsoleLogItem;
-import us.ihmc.scs2.session.mcap.MCAPConsoleLogManager.MCAPLogLevel;
 import us.ihmc.scs2.session.mcap.MCAPLogFileReader;
 import us.ihmc.scs2.session.mcap.MCAPLogSession;
 import us.ihmc.scs2.sessionVisualizer.jfx.SessionVisualizerIOTools;
@@ -45,15 +40,8 @@ import us.ihmc.scs2.sessionVisualizer.jfx.session.log.LogSessionManagerControlle
 import us.ihmc.scs2.sessionVisualizer.jfx.tools.JavaFXMissingTools;
 import us.ihmc.scs2.sessionVisualizer.jfx.tools.PositiveIntegerValueFilter;
 import us.ihmc.scs2.sharedMemory.interfaces.YoBufferPropertiesReadOnly;
-import us.ihmc.scs2.simulation.SpyList;
-import us.ihmc.yoVariables.listener.YoVariableChangedListener;
-import us.ihmc.yoVariables.variable.YoLong;
 
 import java.io.File;
-import java.time.Instant;
-import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
-import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Consumer;
@@ -88,11 +76,12 @@ public class MCAPLogSessionManagerController implements SessionControlsControlle
    @FXML
    private FlowPane videoThumbnailPane;
    @FXML
-   private ListView<MCAPConsoleLogItem> consoleOutputListView;
+   private MCAPConsoleLogOutputPaneController consoleOutputPaneController;
    private Stage stage;
    private SessionVisualizerTopics topics;
    private JavaFXMessager messager;
    private BackgroundExecutorManager backgroundExecutorManager;
+
    private File defaultRobotModelFile = null;
 
    private static String getDate(String filename)
@@ -269,6 +258,8 @@ public class MCAPLogSessionManagerController implements SessionControlsControlle
          logPositionSlider.setDisable(m);
       });
 
+      consoleOutputPaneController.initialize(toolkit);
+
       stage.setScene(new Scene(mainPane));
       stage.setTitle("MCAP Log session controls");
       stage.getIcons().add(SessionVisualizerIOTools.LOG_SESSION_IMAGE);
@@ -301,16 +292,7 @@ public class MCAPLogSessionManagerController implements SessionControlsControlle
       thumbnailsTitledPane.setExpanded(logHasVideos);
       thumbnailsTitledPane.setDisable(!logHasVideos);
 
-      // Setup the console output
-      consoleOutputListView.setCellFactory(param -> new MCAPConsoleLogItemListCell(session.getMCAPLogFileReader().getCurrentTimestamp()));
-      consoleOutputListView.getItems().clear();
-      SpyList<MCAPConsoleLogItem> sessionLogItems = session.getMCAPLogFileReader().getConsoleLogManager().getAllConsoleLogItems();
-      consoleOutputListView.getItems().setAll(sessionLogItems);
-      sessionLogItems.addListener((change) ->
-                                  {
-                                     if (change.wasAdded())
-                                        JavaFXMissingTools.runLater(getClass(), () -> consoleOutputListView.getItems().setAll(sessionLogItems));
-                                  });
+      consoleOutputPaneController.startSession(session);
 
       JavaFXMissingTools.runNFramesLater(5, () -> stage.sizeToScene());
       JavaFXMissingTools.runNFramesLater(6, () -> stage.toFront());
@@ -324,7 +306,7 @@ public class MCAPLogSessionManagerController implements SessionControlsControlle
       endSessionButton.setDisable(true);
       logPositionSlider.setDisable(true);
       multiVideoViewerObjectProperty.set(null);
-      consoleOutputListView.getItems().clear();
+      consoleOutputPaneController.stopSession();
    }
 
    public void openLogFile()
@@ -439,87 +421,5 @@ public class MCAPLogSessionManagerController implements SessionControlsControlle
          }
       };
       messager.addFXTopicListener(topics.getSessionRobotDefinitionListChangeState(), listener);
-   }
-
-   private static class MCAPConsoleLogItemListCell extends javafx.scene.control.ListCell<MCAPConsoleLogItem>
-   {
-      private final Color defaultColor = Color.BLACK;
-      private final Map<MCAPLogLevel, Color> logLevelToColorMap = Map.of(MCAPLogLevel.UNKNOWN,
-                                                                         defaultColor,
-                                                                         MCAPLogLevel.INFO,
-                                                                         Color.CORNFLOWERBLUE,
-                                                                         MCAPLogLevel.WARNING,
-                                                                         Color.ORANGE,
-                                                                         MCAPLogLevel.ERROR,
-                                                                         Color.RED,
-                                                                         MCAPLogLevel.FATAL,
-                                                                         Color.DARKRED);
-
-      private final Color futureColor = Color.GRAY.deriveColor(0, 1, 1, 0.5);
-      private final Map<MCAPLogLevel, String> logLevelToStringMap = Map.of(MCAPLogLevel.UNKNOWN,
-                                                                           "  ???",
-                                                                           MCAPLogLevel.INFO,
-                                                                           " INFO",
-                                                                           MCAPLogLevel.WARNING,
-                                                                           " WARN",
-                                                                           MCAPLogLevel.ERROR,
-                                                                           "ERROR",
-                                                                           MCAPLogLevel.FATAL,
-                                                                           "FATAL");
-      private final DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("HH:mm:ss.SSS a z");
-      private final ZoneId zoneId = ZoneId.systemDefault(); // Need to parameterize this.
-      private final YoLong currentTimestamp;
-      private YoVariableChangedListener timestampListener;
-
-      public MCAPConsoleLogItemListCell(YoLong currentTimestamp)
-      {
-         this.currentTimestamp = currentTimestamp;
-      }
-
-      @Override
-      protected void updateItem(MCAPConsoleLogItem item, boolean empty)
-      {
-         super.updateItem(item, empty);
-
-         if (empty || item == null)
-         {
-            setText(null);
-            setGraphic(null);
-            if (timestampListener != null)
-               currentTimestamp.removeListener(timestampListener);
-            timestampListener = null;
-         }
-         else
-         {
-            setFont(Font.font("Monospaced", 14.0));
-
-            updateTextFill(item);
-            timestampListener = v -> updateTextFill(item);
-            currentTimestamp.addListener(timestampListener);
-            String dateTimeFormatted = dateTimeFormatter.format(item.instant().atZone(zoneId));
-            setText("[%s] [%s]\n\t[%s]: %s".formatted(logLevelToStringMap.get(item.logLevel()), dateTimeFormatted, item.processName(), item.message()));
-            setGraphic(null);
-         }
-      }
-
-      private void updateTextFill(MCAPConsoleLogItem item)
-      {
-         if (item.logTime() > currentTimestamp.getValue())
-            setTextFill(futureColor);
-         else if (logLevelToColorMap.containsKey(item.logLevel()))
-            setTextFill(logLevelToColorMap.get(item.logLevel()));
-         else
-            setTextFill(logLevelToColorMap.get(MCAPLogLevel.UNKNOWN));
-      }
-   }
-
-   public static void main(String[] args)
-   {
-      int seconds = 1701192801;
-      int nanos = 789030589;
-      Instant instant = Instant.ofEpochSecond(seconds, nanos);
-      System.out.println(instant);
-      DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("HH:mm:ss.SSS a z");
-      System.out.println(dateTimeFormatter.format(instant.atZone(ZoneId.systemDefault())));
    }
 }
